# openapi.yaml
openapi: 3.0.3
info:
  title: LeanAutoLinks REST API
  version: 1.0.0
  description: |
    The LeanAutoLinks REST API provides programmatic access to all plugin functionality:
    managing linking rules, monitoring the background processing queue, viewing applied
    links and statistics, managing exclusions, and monitoring performance and cache health.

    ## Authentication

    All endpoints require WordPress Application Passwords using HTTP Basic Auth.
    The authenticated user must have the `manage_options` capability.

    ```bash
    curl -u "admin:xxxx xxxx xxxx xxxx xxxx xxxx" \
      https://example.com/wp-json/leanautolinks/v1/rules
    ```

    Generate an Application Password at **Users > Profile > Application Passwords** in
    your WordPress admin dashboard.

    ## Pagination

    List endpoints follow WordPress REST API conventions. Use `page` and `per_page`
    query parameters to paginate results. Responses include the following headers:

    | Header | Description |
    |--------|-------------|
    | `X-WP-Total` | Total number of items matching the query |
    | `X-WP-TotalPages` | Total number of pages |

    ## Error Handling

    All error responses use a consistent JSON structure:

    ```json
    {
      "code": "invalid_parameter",
      "message": "The rule_type parameter must be one of: internal, affiliate, entity.",
      "data": { "status": 400, "param": "rule_type" }
    }
    ```

    ## Rate Limiting

...

    ## Versioning

...
  contact:
...
servers:
...
security:
  - basicAuth: []
...
paths:
...
  /queue:
    get:
      operationId: listQueue
      summary: List queue items
      description: |
        Returns a paginated list of items in the processing queue. Use the
        `status` filter to view only pending, processing, done, or failed items.
        Results are ordered by scheduled time (ascending), so higher-priority
        items appear first.
      tags: [Queue]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: status
          in: query
          description: Filter by queue item status.
          required: false
          schema:
            $ref: '#/components/schemas/QueueStatus'
        - name: triggered_by
          in: query
          description: Filter by the event that triggered the queue entry.
          required: false
          schema:
            type: string
            enum: [save_post, rule_change, bulk_reprocess, manual]
        - name: orderby
          in: query
          description: Field to order results by.
          required: false
          schema:
            type: string
            enum: [id, post_id, status, scheduled_at, processed_at]
            default: scheduled_at
        - name: order
          in: query
          description: Sort direction.
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: A paginated list of queue items.
          headers:
            X-WP-Total:
              $ref: '#/components/headers/X-WP-Total'
            X-WP-TotalPages:
              $ref: '#/components/headers/X-WP-TotalPages'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QueueItem'
              example:
                - id: 501
                  post_id: 8234
                  status: pending
                  triggered_by: save_post
                  attempts: 0
                  scheduled_at: "2026-03-24T10:15:00"
                  processed_at: null
                  error_log: null
                - id: 499
                  post_id: 7012
                  status: done
                  triggered_by: rule_change
                  attempts: 1
                  scheduled_at: "2026-03-24T10:10:00"
                  processed_at: "2026-03-24T10:10:03"
                  error_log: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /queue/bulk:
    post:
      operationId: triggerBulkReprocess
      summary: Trigger bulk reprocessing
      description: |
        Enqueues posts for background reprocessing. You can reprocess all
        published posts, filter by post type, or specify a date range. Bulk
        jobs run at the lowest priority (50) so they do not block processing
        of newly published or updated posts.

        This operation is idempotent: posts already in the queue with a
        `pending` or `processing` status are not re-enqueued.
      tags: [Queue]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filter]
              properties:
                filter:
                  type: string
                  description: |
                    The scope of posts to reprocess.
                    - `all`: All published posts.
                    - `post_type`: Posts of a specific type (requires `post_type` field).
                    - `date_range`: Posts within a date range (requires `date_from` and/or `date_to`).
                  enum: [all, post_type, date_range]
                post_type:
                  type: string
                  description: |
                    WordPress post type to filter by. Required when `filter` is `post_type`.
                  example: post
                date_from:
                  type: string
                  format: date
                  description: |
                    Start date (inclusive) for the date range filter. Format: `YYYY-MM-DD`.
                    Required when `filter` is `date_range` (at least one of `date_from` or `date_to`).
                  example: "2026-01-01"
                date_to:
                  type: string
                  format: date
                  description: |
                    End date (inclusive) for the date range filter. Format: `YYYY-MM-DD`.
                  example: "2026-03-24"
            examples:
              all_posts:
                summary: Reprocess all posts
                value:
                  filter: all
              by_post_type:
                summary: Reprocess only glossary CPT
                value:
                  filter: post_type
                  post_type: glossary
              by_date_range:
                summary: Reprocess posts from last month
                value:
                  filter: date_range
                  date_from: "2026-02-24"
                  date_to: "2026-03-24"
      responses:
        '202':
          description: Bulk reprocessing enqueued successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  enqueued:
                    type: integer
                    description: Number of posts enqueued for reprocessing.
                    example: 15230
                  estimated_minutes:
                    type: number
                    description: Estimated time to complete based on current throughput.
                    example: 42.3
                  message:
                    type: string
                    example: "15230 posts enqueued for reprocessing at priority 50."
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /queue/retry:
    post:
      operationId: retryFailedItems
      summary: Retry failed queue items
      description: |
        Resets all items with `failed` status back to `pending` with their
        attempt counter reset to 0. These items re-enter the processing queue
        and are picked up by the next Action Scheduler run.
      tags: [Queue]
      responses:
        '200':
          description: Failed items re-enqueued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  retried:
                    type: integer
                    description: Number of failed items moved back to pending.
                    example: 7
                  message:
                    type: string
                    example: "7 failed items re-enqueued for processing."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /queue/clear-done:
    delete:
      operationId: clearDoneItems
      summary: Clear completed queue items
      description: |
        Removes all items with `done` status from the queue table. This is a
        cleanup operation that reduces table size. It does not affect pending,
        processing, or failed items.
      tags: [Queue]
      responses:
        '200':
          description: Completed items cleared.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cleared:
                    type: integer
                    description: Number of completed items removed from the queue.
                    example: 14832
                  message:
                    type: string
                    example: "14832 completed items removed from the queue."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        WordPress Application Passwords. The username is your WordPress
        username and the password is an Application Password generated at
        **Users > Profile > Application Passwords**.

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      description: Current page of the collection.
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      description: Maximum number of items to return per page.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  headers:
    X-WP-Total:
      description: Total number of items in the collection.
      schema:
        type: integer
    X-WP-TotalPages:
      description: Total number of pages in the collection.
      schema:
        type: integer

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    QueueStatus:
      type: string
      enum: [pending, processing, done, failed]
      description: |
        The processing status of a queue item:
        - `pending`: Waiting to be picked up by Action Scheduler.
        - `processing`: Currently being processed by a background worker.
        - `done`: Successfully processed and cached.
        - `failed`: Processing failed after maximum retry attempts (5).
...
    QueueItem:
      type: object
      description: An item in the background processing queue.
      required:
        - id
        - post_id
        - status
        - triggered_by
        - attempts
        - scheduled_at
      properties:
        id:
          type: integer
          format: int64
          description: Unique queue entry identifier.
          readOnly: true
          example: 501
        post_id:
          type: integer
          format: int64
          description: The WordPress post ID to be processed.
          example: 8234
        status:
          $ref: '#/components/schemas/QueueStatus'
        triggered_by:
          type: string
          maxLength: 50
          description: |
            What triggered this queue entry:
            - `save_post`: Post was created or updated.
            - `rule_change`: A rule was created, updated, or deleted.
            - `bulk_reprocess`: Triggered via the bulk reprocess endpoint.
            - `manual`: Manually enqueued via admin UI or WP-CLI.
          enum: [save_post, rule_change, bulk_reprocess, manual]
          default: save_post
          example: save_post
        attempts:
          type: integer
          minimum: 0
          maximum: 5
          description: |
            Number of processing attempts. After 5 failed attempts, the item
            is marked as `failed` and no further retries are scheduled
            automatically.
          default: 0
          example: 0
        scheduled_at:
          type: string
          format: date-time
          description: When the item was enqueued for processing.
          example: "2026-03-24T10:15:00"
        processed_at:
          type: string
          format: date-time
          nullable: true
          description: When the item was successfully processed. Null if not yet processed.
          example: null
        error_log:
          type: string
          nullable: true
          description: Error message from the last failed attempt. Null if no errors.
          example: null

    # --- Applied Links -------------------------------------------------------
...
    Error:
      type: object
      description: |
        Standard error response following WordPress REST API conventions.
      required:
        - code
        - message
        - data
      properties:
        code:
          type: string
          description: |
            Machine-readable error code. Common codes:
            - `rest_forbidden`: Missing required capability.
            - `rest_no_route`: Endpoint not found.
            - `invalid_parameter`: A request parameter failed validation.
            - `missing_parameter`: A required parameter is missing.
            - `not_found`: The requested resource does not exist.
            - `duplicate_keyword`: A rule with this keyword already exists.
            - `internal_error`: An unexpected server error occurred.
          example: invalid_parameter
        message:
          type: string
          description: Human-readable error description.
          example: "The rule_type parameter must be one of: internal, affiliate, entity."
        data:
          type: object
          description: Additional error context.
          properties:
            status:
              type: integer
              description: The HTTP status code.
              example: 400
            param:
              type: string
              description: The parameter that caused the error (if applicable).
              example: rule_type
            details:
              type: object
              description: Additional error details (structure varies by error type).

    # --- Pagination ----------------------------------------------------------
...
  responses:
    BadRequest:
      description: |
        The request is malformed or contains invalid parameters. Check the
        `code` and `data.param` fields to identify the issue.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: invalid_parameter
            message: "The rule_type parameter must be one of: internal, affiliate, entity."
            data:
              status: 400
              param: rule_type
    Unauthorized:
      description: |
        Authentication is missing or invalid. Ensure you are sending a valid
        WordPress Application Password via HTTP Basic Auth.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: rest_not_logged_in
            message: "You are not currently logged in."
            data:
              status: 401
    Forbidden:
      description: |
        The authenticated user does not have the `manage_options` capability
        required to access this endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: rest_forbidden
            message: "Sorry, you are not allowed to manage LeanAutoLinks settings."
            data:
              status: 403
...
    UnprocessableEntity:
      description: |
        The request body is syntactically valid but contains semantic errors.
        For example, a required field is missing or a value violates business rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: missing_parameter
            message: "The keyword field is required."
            data:
              status: 422
              param: keyword
    InternalError:
      description: An unexpected server error occurred. Check the WordPress error log for details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: internal_error
            message: "An unexpected error occurred. Please try again later."
            data:
              status: 500